2010-10-05  Ole Reinhardt <ole.reinhardt@embedded-it.de>
	* tcpsm.c:  
	
	  Fixed bug #3080997 in nut/net/tcpsm.c with fix from Zack which was
	  suggested just in 2007!!!

	  Thanks Zack and Harald for digging out this solution again!
	            


git-svn-id: svn+ssh://svn.code.sf.net/p/ethernut/code/trunk@3147 1ca430b3-8271-4a6b-9b0e-3dcac98a8924

+++ nut/net/tcpsm.c
@@ -1708,10 +1708,11 @@ THREAD(NutTcpSm, arg)
                  * Process retransmit timer.
                  */
                 if (sock->so_tx_nbq && sock->so_retran_time) {
-                    if ((uint16_t)NutGetMillis() - (sock->so_retran_time & ~1) >= sock->so_rtto) {
+                    if ((uint16_t)((uint16_t)NutGetMillis() - (sock->so_retran_time & ~1)) >= sock->so_rtto) {
                         NutTcpStateRetranTimeout(sock);
                     }
                 }
+
 
                 /*
                  * Destroy sockets after timeout in TIMEWAIT state.
